api-server/requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
pymysql==0.10.0
dbutils==1.3
pyjwt==1.7.1